# src/main/java/com/moviebooking/dto/response/AuthResponse.java
package com.moviebooking.dto.response;

import com.moviebooking.enums.Role;

public class AuthResponse {

    private String token;
    private String email;
    private Role role;

    public AuthResponse() {}

    public AuthResponse(String token, String email, Role role) {
        this.token = token;
        this.email = email;
        this.role = role;
    }

    public String getToken() { return token; }
    public void setToken(String token) { this.token = token; }
    public String getEmail() { return email; }
    public void setEmail(String email) { this.email = email; }
    public Role getRole() { return role; }
    public void setRole(Role role) { this.role = role; }
}