# swagger.yaml
openapi: 3.0.3
info:
  title: Movie Booking System API
  description: |
    Multi-city, multi-theatre movie ticket booking platform built on Java 21 + Spring Boot 3.3.5.

    ## Authentication
    Most endpoints require a JWT Bearer token. Obtain one via `POST /api/auth/login` or `POST /api/auth/register`.
    Pass it in the `Authorization` header as `Bearer <token>`.

    ## Roles
    | Role | Access |
    |---|---|
    | **CUSTOMER** | Browse, hold seats, book, cancel, view history |
    | **ADMIN** | Full catalogue management (cities, theatres, screens, movies, shows, pricing, discounts, refund policies) |

    ## Error Shape
    All error responses follow `{ "status": <int>, "message": "<string>", "timestamp": "<ISO-8601>" }`.
  version: 1.0.0
  contact:
    email: manya@adobe.com
servers:
  - url: http://localhost:8080
    description: Local development server
...
security: []   # default: public; overridden per-operation
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  # ─── Enums ──────────────────────────────────────────────────────────────────
  schemas:
    Role:
      type: string
      enum: [ADMIN, CUSTOMER]
    SeatType:
      type: string
      enum: [REGULAR, PREMIUM]
...
    SeatStatus:
      type: string
      enum: [AVAILABLE, HELD, BOOKED]
    BookingStatus:
      type: string
      enum: [PENDING, CONFIRMED, CANCELLED, REFUNDED]
      description: PENDING is reserved for a future async payment flow.
...
    PaymentMethod:
      type: string
      enum: [UPI, CARD, WALLET]
...
    RegisterRequest:
      type: object
      required: [name, email, password]
      properties:
        name:
          type: string
          example: Alice Smith
        email:
          type: string
          format: email
          example: alice@example.com
        password:
          type: string
          minLength: 6
          example: secret123
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          example: alice@example.com
        password:
          type: string
          example: secret123
...
    CreateMovieRequest:
      type: object
      required: [title]
      properties:
        title:
          type: string
          example: Interstellar
        description:
          type: string
          maxLength: 1000
          example: A team of explorers travel through a wormhole in space.
        duration:
          type: integer
          minimum: 1
          description: Duration in minutes
          example: 169
        language:
          type: string
          example: English
        genre:
          type: string
          example: Sci-Fi
...
    CreateSeatHoldRequest:
      type: object
      required: [showId, seatIds]
      properties:
        showId:
          type: integer
          format: int64
          example: 5
        seatIds:
          type: array
          items:
            type: integer
            format: int64
          minItems: 1
          example: [10, 11, 12]
    CreateBookingRequest:
      type: object
      required: [holdId, paymentMethod]
      properties:
        holdId:
          type: integer
          format: int64
          example: 7
        paymentMethod:
          $ref: '#/components/schemas/PaymentMethod'
        discountCode:
          type: string
          nullable: true
          example: SUMMER20

    # ─── Response schemas ──────────────────────────────────────────────────────
    AuthResponse:
      type: object
      properties:
        token:
          type: string
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        email:
          type: string
          format: email
          example: alice@example.com
        role:
          $ref: '#/components/schemas/Role'
...
    ShowSeatResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 101
        seatId:
          type: integer
          format: int64
          example: 10
        rowNumber:
          type: integer
          example: 3
        seatNumber:
          type: integer
          example: 12
        seatType:
          $ref: '#/components/schemas/SeatType'
        status:
          $ref: '#/components/schemas/SeatStatus'
    MovieResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        title:
          type: string
          example: Interstellar
        description:
          type: string
          example: A team of explorers travel through a wormhole in space.
        duration:
          type: integer
          description: Duration in minutes
          example: 169
        language:
          type: string
          example: English
        genre:
          type: string
          example: Sci-Fi
    ShowResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 5
        movieId:
          type: integer
          format: int64
          example: 1
        movieTitle:
          type: string
          example: Interstellar
        screenId:
          type: integer
          format: int64
          example: 3
        screenName:
          type: string
          example: Screen 1
        theatreName:
          type: string
          example: PVR Cinemas
        startTime:
          type: string
          format: date-time
          example: "2025-07-15T18:00:00"
        endTime:
          type: string
          format: date-time
          example: "2025-07-15T20:49:00"
    SeatHoldResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 7
        showId:
          type: integer
          format: int64
          example: 5
        showSeatIds:
          type: array
          items:
            type: integer
            format: int64
          example: [101, 102, 103]
        createdAt:
          type: string
          format: date-time
          example: "2025-07-15T17:30:00"
        expiresAt:
          type: string
          format: date-time
          example: "2025-07-15T17:40:00"
    BookingResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 20
        showId:
          type: integer
          format: int64
          example: 5
        movieTitle:
          type: string
          example: Interstellar
        bookedSeats:
          type: array
          items:
            $ref: '#/components/schemas/ShowSeatResponse'
        totalAmount:
          type: number
          format: decimal
          example: 450.00
        bookingStatus:
          $ref: '#/components/schemas/BookingStatus'
        bookedAt:
          type: string
          format: date-time
          example: "2025-07-15T17:32:00"
...
    ErrorResponse:
      type: object
      properties:
        status:
          type: integer
          example: 400
        message:
          type: string
          example: Seat is not available for booking
        timestamp:
          type: string
          format: date-time
          example: "2025-07-15T17:30:00"

    # ─── Paginated wrappers ────────────────────────────────────────────────────
    PageMetadata:
      type: object
      properties:
        totalElements:
          type: integer
          format: int64
          example: 50
        totalPages:
          type: integer
          example: 3
        size:
          type: integer
          example: 20
        number:
          type: integer
          description: Zero-based current page index
          example: 0
        first:
          type: boolean
          example: true
        last:
          type: boolean
          example: false
        numberOfElements:
          type: integer
          example: 20
        empty:
          type: boolean
          example: false
    PagedMovieResponse:
      allOf:
        - $ref: '#/components/schemas/PageMetadata'
        - type: object
          properties:
            content:
              type: array
              items:
                $ref: '#/components/schemas/MovieResponse'
    PagedShowResponse:
      allOf:
        - $ref: '#/components/schemas/PageMetadata'
        - type: object
          properties:
            content:
              type: array
              items:
                $ref: '#/components/schemas/ShowResponse'
    PagedBookingResponse:
      allOf:
        - $ref: '#/components/schemas/PageMetadata'
        - type: object
          properties:
            content:
              type: array
              items:
                $ref: '#/components/schemas/BookingResponse'

  # ─── Reusable responses ──────────────────────────────────────────────────────
  responses:
    Unauthorized:
      description: Missing or invalid JWT token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            status: 401
            message: Unauthorized
            timestamp: "2025-07-15T17:30:00"
    Forbidden:
      description: Authenticated but role not permitted
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            status: 403
            message: Access denied
            timestamp: "2025-07-15T17:30:00"
    NotFound:
      description: Requested resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            status: 404
            message: Resource not found
            timestamp: "2025-07-15T17:30:00"
    Conflict:
      description: Optimistic lock conflict — seat was modified concurrently; retry the request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            status: 409
            message: Seat was modified by another transaction; please retry
            timestamp: "2025-07-15T17:30:00"

  # ─── Reusable parameters ─────────────────────────────────────────────────────
  parameters:
    PageParam:
      name: page
      in: query
      description: Zero-based page index (default 0)
      schema:
        type: integer
        default: 0
        minimum: 0
    SizeParam:
      name: size
      in: query
      description: Page size (default 20)
      schema:
        type: integer
        default: 20
        minimum: 1

# ═══════════════════════════════════════════════════════════════════════════════
# Paths
# ═══════════════════════════════════════════════════════════════════════════════
paths:
...
  /api/auth/register:
    post:
      tags: [Auth]
      summary: Register a new user
      operationId: register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '200':
          description: Registration successful — returns JWT token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Validation error or email already in use
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/auth/login:
    post:
      tags: [Auth]
      summary: Authenticate and receive a JWT token
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful — returns JWT token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Bad credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ─── Movies (public) ───────────────────────────────────────────────────────
  /api/movies:
    get:
      tags: [Movies]
      summary: List all movies (paginated)
      operationId: getAllMovies
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
      responses:
        '200':
          description: Paginated list of movies
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagedMovieResponse'
  /api/movies/{id}:
    get:
      tags: [Movies]
      summary: Get a movie by ID
      operationId: getMovieById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Movie detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MovieResponse'
        '404':
          $ref: '#/components/responses/NotFound'

  # ─── Shows (public) ────────────────────────────────────────────────────────
  /api/cities/{cityId}/shows:
    get:
      tags: [Shows]
      summary: List all shows in a city (paginated)
      operationId: getShowsByCity
      parameters:
        - name: cityId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
      responses:
        '200':
          description: Paginated list of shows in the city
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagedShowResponse'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/shows/{id}:
    get:
      tags: [Shows]
      summary: Get show detail by ID
      operationId: getShowById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Show detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ShowResponse'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/shows/{id}/seats:
    get:
      tags: [Shows]
      summary: Get seat availability for a show
      operationId: getShowSeats
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: List of show seats with current availability status
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ShowSeatResponse'
        '404':
          $ref: '#/components/responses/NotFound'

  # ─── Seat Holds (CUSTOMER) ─────────────────────────────────────────────────
  /api/holds:
    post:
      tags: [Holds]
      summary: Hold one or more seats for a show
      description: |
        Temporarily reserves the requested seats for the authenticated customer.
        The hold expires after the configured hold duration (default 10 minutes).
        All seats must be AVAILABLE; if any seat is HELD or BOOKED the request is rejected.
        Returns HTTP 409 if an optimistic lock conflict is detected (concurrent selection of the same seat).
      operationId: createHold
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSeatHoldRequest'
      responses:
        '200':
          description: Hold created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SeatHoldResponse'
        '400':
          description: One or more seats are not available or not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'

  # ─── Bookings (CUSTOMER) ───────────────────────────────────────────────────
  /api/bookings:
    post:
      tags: [Bookings]
      summary: Confirm a seat hold into a booking
      description: |
        Converts an active hold into a confirmed booking by processing payment.

        **Happy path:** validates hold ownership and expiry → calculates price per seat from
        PricingTier (SeatType × DayType) → applies optional discount → processes payment via
        the selected gateway → marks seats BOOKED → creates Booking (CONFIRMED) + Payment (SUCCESS)
        → fires async `BookingConfirmedEvent`.

        **Payment failure:** resets all held seats to AVAILABLE and expires the hold immediately;
        does not create a Booking record. Returns HTTP 400.
      operationId: createBooking
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBookingRequest'
      responses:
        '200':
          description: Booking confirmed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookingResponse'
        '400':
          description: |
            One of:
            - Hold has expired
            - Hold does not belong to the authenticated customer
            - Payment gateway returned failure
            - Discount code invalid, expired, or over usage limit
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: Hold not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/bookings/{id}/cancel:
    patch:
      tags: [Bookings]
      summary: Cancel a confirmed booking
      description: |
        Cancels a CONFIRMED booking. Refund amount is determined by the closest matching
        RefundPolicy time band (`minHoursBeforeShow`).

        - `refundAmount > 0` → `bookingStatus = REFUNDED`; fires `BookingCancelledEvent` + `RefundProcessedEvent`
        - `refundAmount = 0` → `bookingStatus = CANCELLED`; fires `BookingCancelledEvent` only

        Cancelled seats are immediately returned to AVAILABLE status.
      operationId: cancelBooking
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Booking cancelled (or refunded)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookingResponse'
        '400':
          description: |
            One of:
            - Booking does not belong to the authenticated customer
            - Booking is not in CONFIRMED status
            - Show has already started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/bookings/history:
    get:
      tags: [Bookings]
      summary: Get booking history for the authenticated customer (paginated)
      operationId: getBookingHistory
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
      responses:
        '200':
          description: Paginated booking history
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagedBookingResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ─── Admin — Cities ────────────────────────────────────────────────────────
...
  /api/admin/movies/{id}:
    put:
      tags: [Admin — Movies]
      summary: Update a movie
      operationId: updateMovie
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateMovieRequest'
      responses:
        '200':
          description: Movie updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MovieResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Admin — Movies]
      summary: Delete a movie
      operationId: deleteMovie
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Movie deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # ─── Admin — Shows ─────────────────────────────────────────────────────────
...